Create Bombadier Vulture Mod
Original Mod Steam Description: "This mod adds a completely new vulture with the ability to basically drop bombs.
This species of vulture of unknown origin seems to be coated in an explosive substance which accumulates and solidifies in its lowers part until it falls in the form of a 'seed boom', these seeds are highly dangerous since once they touch the ground any contact like hitting them or colliding with them will light them up and soon explode. Due to this 'destructive hunting method', the bomber vulture prefers to live in large esplanades to maximize its kills. If you see something fall from the sky in the shape of a seed, look for cover and fast...
If anyone wants to put it in a custom region, don't be shy about asking me.
I hope you like it!"